support itemRef weights with floating points

+++ Scoring/Model/AssessmentItemRef.cs
@@ -6,7 +6,7 @@ namespace Citolab.QTI.ScoringEngine.Model
     internal class AssessmentItemRef
     {
         public string Identifier { get; set; }
-        public Dictionary<string, int> Weights { get; set; }
+        public Dictionary<string, float> Weights { get; set; }
         public HashSet<string> Categories { get; set; }
     }
 }
